requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
Flask-Mail==0.10.0
Flask-Session==0.6.0
groq==0.4.1
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
email-validator==2.1.1
alembic==1.13.2
uuid==1.30
Werkzeug==3.0.3

